* milter/manager/milter-manager-main.c: Do not kill if launcher_pid is 	not valid.

+++ milter/manager/milter-manager-main.c
@@ -131,7 +131,8 @@ shutdown_client (int signum)
 {
     if (current_client)
         milter_client_shutdown(current_client);
-    kill(launcher_pid, SIGKILL);
+    if (launcher_pid > 0)
+        kill(launcher_pid, SIGKILL);
 }
 
 static void
